% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdifdif.R
\docType{package}
\name{cdifdif-package}
\alias{cdifdif-package}
\alias{_PACKAGE}
\title{cdifdif: Spillover Robust Difference-in-Differences Estimation}
\description{
This package implements spillover robust difference-in-differences
estimates as described in Clarke (2017).  It tests for local spillovers in a
diff-in-diff setting, and optimally determines spillover propagation.
It returns estimates of both the treatment effect, as well as (if relevant)
the treatment effect on those close to treated, who weren't themselves treated.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/damiancclarke/cdifdif}
  \item Report bugs at \url{https://github.com/damiancclarke/cdifdif/issues}
}

}
\author{
\strong{Maintainer}: Damian Clarke \email{damian.c.clarke@gmail.com}

Authors:
\itemize{
  \item Joshua Kunst \email{jbkunst@gmail.com}
}

}
