% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdifdif-algo.R
\name{cdifdif}
\alias{cdifdif}
\title{Spillover Robust Diff-in-Diff Estimation}
\usage{
cdifdif(formula, data, dist, maxDist = 30, delta = 1, alpha = 0.05,
  k = 1000, verbose = TRUE)
}
\arguments{
\item{formula}{formula.}

\item{data}{data.}

\item{dist}{a N-by-1 distance to treatment.}

\item{maxDist}{a maximum spillover bandwidth to consider.}

\item{delta}{a step-size for bandwidth search (based on dist variable).}

\item{alpha}{a minimum t-stat to consider marginal spillover to be significant.}

\item{k}{a number of folds for k-fold Cross-Validation.}

\item{verbose}{Verbose}
}
\description{
Implements Spillover Robust Diff-in-Diff Estimates (Clarke, 2017)(Clarke, 2017)
}
