% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdifdif-algo.R
\name{data}
\alias{data}
\title{Spillover Robust Diff-in-Diff Estimation}
\usage{
data(y, X, dist, maxDist, delta, tlimit, CVtype, kfolds)
}
\arguments{
\item{y}{a N-by-1 dependent variable.}

\item{X}{a N-by-k baseline independent variables}

\item{dist}{a N-by-1 distance to treatment.}

\item{maxDist}{a maximum spillover bandwidth to consider.}

\item{delta}{a step-size for bandwidth search (based on dist variable).}

\item{tlimit}{a minimum t-stat to consider marginal spillover to be significant.}

\item{CVtype}{a type of Cross-Validation (must be either 'kfoldcv' or 'loocv').}

\item{kfolds}{a number of folds for k-fold Cross-Validation.}
}
\description{
Implements Spillover Robust Diff-in-Diff Estimates (Clarke, 2017)(Clarke, 2017)
}
