% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spilloverDGP}
\alias{spilloverDGP}
\title{spilloverDGP data example}
\format{A \code{data frame} with 1,000 observations and 8 variables.}
\usage{
spilloverDGP
}
\description{
Example data for demo.
}
\section{Variables}{


\itemize{

\item \code{id}: The months.
\item \code{dist}: Tokyo's temperatures.
\item \code{x}: New York's temperatures.
\item \code{time}: Berlin's temperatures.
\item \code{treat}: London's temperatures.
\item \code{y1}: London's temperatures.
\item \code{y2}: London's temperatures.
\item \code{y3}: London's temperatures.

}
}

\keyword{datasets}
